Move RewriteRecipeLauncher to top-level package
