+++ spring-rewrite-commons-launcher/src/main/java/org/springframework/rewrite/boot/autoconfigure/RewriteLauncherConfiguration.java
@@ -18,7 +18,7 @@ package org.springframework.rewrite.boot.autoconfigure;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
-import org.springframework.rewrite.execution.RewriteRecipeLauncher;
+import org.springframework.rewrite.RewriteRecipeLauncher;
 import org.springframework.rewrite.RewriteProjectParser;
 import org.springframework.rewrite.resource.ProjectResourceSetFactory;
 import org.springframework.rewrite.resource.ProjectResourceSetSerializer;
